+++ docs/CHANGELOG.md
@@ -8,6 +8,8 @@
 - Fixed  `ImportError` which was raised when trying to import `dash_uploader` when `packaging` was not installed. [PR 54](https://github.com/np-8/dash-uploader/pull/54)
 ### Changed 
 - WIP: resumable.js -> flowjs [[Issue 21](https://github.com/np-8/dash-uploader/issues/21)]
+- **Backwards incompatible**: Changed the CSS class of the component to be `dash-uploader-completed`,  instead of `dash-uploader-complete`, when upload is completed. 
+
 ## v.0.6.0 (2021-09-19)
 ### Added 
 - New `chunk_size`, `disabled` and `text_disabled` parameters for `du.Upload`. [Issue 41](https://github.com/np-8/dash-uploader/issues/41)